% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_CoViData.R
\name{Get_CoViData}
\alias{Get_CoViData}
\title{Download regional CoViD cases data from the Italian Superior Institute of Health}
\source{
\href{https://github.com/pcm-dpc/COVID-19/tree/master/dati-regioni}{GitHub archive}
}
\usage{
Get_CoViData(
  date_from = "2022-01-01",
  date_to = "2022-01-15",
  verbose = TRUE,
  autoAbort = FALSE
)
}
\arguments{
\item{date_from}{Character. The starting date. Important: necessary to use the "yyyy-mm-dd" format.}

\item{date_to}{Character. The ending date. Important: necessary to use the "yyyy-mm-dd" format.}

\item{verbose}{Logical. If \code{TRUE}, the user keeps track of the main underlying operations. \code{TRUE} by default.}

\item{autoAbort}{Logical. Whether to automatically abort the operation and return NULL in case of missing internet connection or server response errors. \code{FALSE} by default.}
}
\value{
An object of class \code{tbl_df}, \code{tbl} and \code{data.frame}.
}
\description{
This function downloads the region-level CoViD cases for a user-specified time period
}
\details{
Covid cases data are provided by the Italian Superior Institute of Health and hosted on GitHub.
The current package version of this function only downloads the data between two dates.
}
\examples{

\donttest{
  Get_CoViData(date_from = "2022-01-01", date_to = "2022-01-08")

}


}
